EPP-106 abstract notify helpers
